+++ infra/src/main/kotlin/fr/sacane/jmanager/app/README.md
