+++ infra/src/main/kotlin/fr/sacane/jmanager/app/adapters/UserControlAdapter.kt
@@ -1,10 +1,10 @@
-package fr.sacane.jmanager.infra.api.adapters
+package fr.sacane.jmanager.app.adapters
 
+import fr.sacane.jmanager.app.*
 import fr.sacane.jmanager.domain.hexadoc.LeftAdapter
 import fr.sacane.jmanager.domain.models.Password
 import fr.sacane.jmanager.domain.port.apiside.LoginManager
 import fr.sacane.jmanager.domain.port.apiside.UserRegister
-import fr.sacane.jmanager.infra.api.*
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
