+++ infra/src/main/kotlin/fr/sacane/jmanager/app/controller/UserEndpoint.kt
@@ -1,0 +1,5 @@
+package fr.sacane.jmanager.app.controller
+
+class UserEndpoint {
+
+}
