+++ infra/src/main/kotlin/fr/sacane/jmanager/infra/README.md
