--- infra/src/main/kotlin/fr/sacane/jmanager/infra/api/controller/UserEndpoint.kt
@@ -1,5 +1,0 @@
-package fr.sacane.jmanager.infra.api.controller
-
-class UserEndpoint {
-
-}
