+++ infra/src/main/kotlin/fr/sacane/jmanager/infra/entity/AccountResource.kt
@@ -1,4 +1,4 @@
-package fr.sacane.jmanager.infra.server.entity
+package fr.sacane.jmanager.infra.entity
 
 import fr.sacane.jmanager.domain.hexadoc.DatasourceEntity
 import org.springframework.lang.Nullable
