+++ infra/src/main/kotlin/fr/sacane/jmanager/infra/repositories/LoginRepository.kt
@@ -1,0 +1,12 @@
+package fr.sacane.jmanager.infra.repositories
+
+import fr.sacane.jmanager.infra.entity.Login
+import fr.sacane.jmanager.infra.entity.UserResource
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+import java.util.UUID
+
+@Repository
+interface LoginRepository: JpaRepository<Login, UUID>{
+    fun findByUser(user: UserResource): Login?
+}
