+++ infra/src/main/kotlin/fr/sacane/jmanager/infra/repositories/SheetRepository.kt
@@ -1,6 +1,6 @@
-package fr.sacane.jmanager.infra.server.repositories
+package fr.sacane.jmanager.infra.repositories
 
-import fr.sacane.jmanager.infra.server.entity.SheetResource
+import fr.sacane.jmanager.infra.entity.SheetResource
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
