+++ infra/src/main/resources/application.properties
@@ -5,8 +5,8 @@ spring.jpa.hibernate.ddl-auto=update
 
 spring.datasource.driverClassName=org.postgresql.Driver
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=##
-spring.datasource.password=##
+spring.datasource.username=postgres
+spring.datasource.password=01012000
 
 spring.jpa.properties.hibernate.globally_quoted_identifiers=true
 spring.jpa.show-sql=true
